st-thief now uses a template which blends the thief over st-bluegray - fixes bug #
Additionally, the graphics for the upcoming fl-thief are drawn
from it, allowing all four kinds of fl-bluegray to be used.
Note however, that the correct assignment of the number of the
model and animation (format "fl-thiefX", "fl-thiefX-emerge",
"fl-thiefX-retreat" and "fl-thiefX-captured" with X=1..4) has
to be done from C++-level, def_randmodel doesn't help much.