Fix bug: Missing replacement when opening oxyds with selection.
When an oxyd is paired with an oxyd which is in its closing-animation,
no replacement by their selection-attributes would be triggered, because
the closing oxyd would not enter the necessary opening-animation.
This is now fixed.

Thanks to Hairball for identifying this bug!

Side effect:

A closing oxyd which is about to open will reach the end of the
opening-animation sooner than a completely closed oxyd would.
Therefore, the replacement will be triggered sooner. If this
behaviour is not wished for, remove the inner if-clause in
OxydStone.cc, set_iState, switch-case OPEN_PAIR.